Move spatialjsonToBinary to tests
